install-windows.bat - check for existing install from within pulled repo
As it is, if a user follows README and downloads install-windows.bat and runs it, it will install StabilitySwarmUI by cloning the repo into the folder `StableSwarmUI`.  This can lead to a recursive issue if the user isn't paying attention - install-windows.bat will exist within the newly cloned directory, and as-is will clone *another* copy into a new StableSwarmUI directory within the existing directory.

Another solution could be to check if `%~dp0` == StabilitySwarmUI (not recommended), or check the .git folder, but this works for now.